Add "Add Card" button to Payment Methods screen
This commit adds an "Add Card" button to the bottom of the Payment Methods screen, allowing users to
 add new payment methods.

The button is implemented using the MyButton widget and is styled with a plus icon and the text "Add Card".

This change improves the usability of the Payment Methods screen by providing a clear and convenient way for users to add new payment methods.